requirements.txt
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
pydantic==2.9.2
fastapi==0.115.4
flake8==7.1.1
black==24.10.0
mypy==1.13.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
uvicorn==0.32.0